commit repo

+++ src/models/AddMaterials.java
@@ -2,7 +2,7 @@ package models;
 
 public class AddMaterials extends SuperClass {
 
-    public AddMaterials(int ID, long counter) {
+    public AddMaterials(int ID, int counter) {
         super(ID, counter);
     }
 }
