+++ src/models/Course.java
@@ -6,7 +6,7 @@ public class Course extends SuperClass{
     private Lectures lectures;
 
 
-    public Course(int ID, long counter) {
+    public Course(int ID, int counter) {
         super(ID, counter);
     }
 }
