+++ src/models/Lectures.java
@@ -5,7 +5,7 @@ public class Lectures extends SuperClass {
     private AddMaterials addMaterials;
     private String name;
 
-    public Lectures(int ID, long counter) {
+    public Lectures(int ID, int counter) {
         super(ID, (int) counter);
     }
