+++ src/models/Students.java
@@ -2,7 +2,7 @@ package models;
 
 public class Students extends SuperClass {
 
-    public Students(int ID, long counter) {
+    public Students(int ID, int counter) {
         super(ID, counter);
     }
 }
