+++ src/models/Teachers.java
@@ -2,7 +2,7 @@ package models;
 
 public class Teachers extends SuperClass {
 
-    public Teachers(int ID, long counter) {
+    public Teachers(int ID, int counter) {
         super(ID, counter);
     }
 }
